src/utils/requirements.txt
pyyaml==5.4.1
pandas==1.1.3
requests